ONYX-43436: Refactor edge endpoint to split to dedicated files
