+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
+## [1.0.6-cloud] - 2023-08-21
+
 ## [1.0.5-cloud] - 2023-08-16
 ### Security
 - Previously, attempting to add and remove a privilege in the same policy load
