+++ app/controllers/edge/api/edge_creation_controller.rb
@@ -1,7 +1,7 @@
 
-require_relative '../controllers/wrappers/policy_wrapper'
+require_relative '../../wrappers/policy_wrapper'
 
-class EdgeCreatorController < RestController
+class EdgeCreationController < RestController
   include AccountValidator
   include BodyParser
   include EdgeValidator
@@ -10,6 +10,30 @@ class EdgeCreatorController < RestController
   include GroupMembershipValidator
   include PolicyWrapper
   include ParamsValidator
+
+  def generate_install_token
+    logger.info(LogMessages::Endpoints::EndpointRequested.new("edge/edge-creds"))
+    allowed_params = %i[account edge_name]
+    options = params.permit(*allowed_params).to_h.symbolize_keys
+    audit_params = { edge_name: options[:edge_name], user: current_user.role_id, client_ip: request.ip }
+    begin
+      validate_conjur_admin_group(options[:account])
+
+      edge = Edge[name: options[:edge_name]] || (raise RecordNotFound.new(options[:edge_name], message: "Edge #{options[:edge_name]} not found"))
+      installer_token = edge.get_installer_token(options[:account], request)
+
+      edge_host_name = Role.username_from_roleid(edge.get_edge_host_name(options[:account]))
+
+    rescue => e
+      audit_params[:error_message] = e.message
+      raise e
+    ensure
+      Audit.logger.log(Audit::Event::CredsGeneration.new(**audit_params))
+    end
+    logger.info(LogMessages::Endpoints::EndpointFinishedSuccessfully.new("edge/edge-creds"))
+    response.set_header("Content-Encoding", "base64")
+    render(plain: Base64.strict_encode64(edge_host_name + ":" + installer_token))
+  end
 
   #this endpoint loads a policy with the edge host values + adds the edge name to Edge table
   def create_edge
